Revert "Move slice_flip_kernel* to separate compilation units (#4637)"
This reverts commit 4f7796e6e395a22a321e29e16c30599ef50a64b2.
