--- dali/kernels/slice/slice_flip_normalize_permute_pad_gpu_u16.cu
@@ -1,28 +1,0 @@
-// Copyright (c) 2023, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include "dali/kernels/slice/slice_flip_normalize_permute_pad_gpu_impl.h"
-
-namespace dali {
-namespace kernels {
-
-CMN_KERNEL_INSTANTIATE_FOREACH_INTYPE(uint16_t, 1);
-CMN_KERNEL_INSTANTIATE_FOREACH_INTYPE(uint16_t, 2);
-CMN_KERNEL_INSTANTIATE_FOREACH_INTYPE(uint16_t, 3);
-CMN_KERNEL_INSTANTIATE_FOREACH_INTYPE(uint16_t, 4);
-CMN_KERNEL_INSTANTIATE_FOREACH_INTYPE(uint16_t, 5);
-CMN_KERNEL_INSTANTIATE_FOREACH_INTYPE(uint16_t, 6);
-
-}  // namespace kernels
-}  // namespace dali
